Update forced layout
