+++ assets/js/blocks/cart-checkout-shared/use-forced-layout/index.ts
@@ -7,93 +7,14 @@ import {
 	createBlock,
 	getBlockType,
 	createBlocksFromInnerBlocksTemplate,
-	BlockInstance,
+	TemplateArray,
 } from '@wordpress/blocks';
-import type { Block, TemplateArray } from '@wordpress/blocks';
-import type { MutableRefObject } from 'react';
-
-interface LockableBlock extends Block {
-	attributes: {
-		lock?: {
-			type: 'object';
-			remove?: boolean;
-			move: boolean;
-			default?: {
-				remove?: boolean;
-				move?: boolean;
-			};
-		};
-	};
-}
-const isBlockLocked = ( {
-	attributes,
-}: {
-	attributes: LockableBlock[ 'attributes' ];
-} ) => Boolean( attributes.lock?.remove || attributes.lock?.default?.remove );
+import { useEditorContext } from '@woocommerce/base-context';
 
 /**
- * This hook is used to determine which blocks are missing from a block. Given the list of inner blocks of a block, we
- * can check for any registered blocks that:
- * a) Are locked,
- * b) Have the parent set as the current block, and
- * c) Are not present in the list of inner blocks.
+ * Internal dependencies
  */
-const getMissingBlocks = (
-	innerBlocks: BlockInstance[],
-	registeredBlockTypes: ( LockableBlock | undefined )[]
-) => {
-	const lockedBlockTypes = registeredBlockTypes.filter(
-		( block: LockableBlock | undefined ) => block && isBlockLocked( block )
-	);
-	const missingBlocks: LockableBlock[] = [];
-	lockedBlockTypes.forEach( ( lockedBlock ) => {
-		if ( typeof lockedBlock === 'undefined' ) {
-			return;
-		}
-		const existingBlock = innerBlocks.find(
-			( block ) => block.name === lockedBlock.name
-		);
-
-		if ( ! existingBlock ) {
-			missingBlocks.push( lockedBlock );
-		}
-	} );
-	return missingBlocks;
-};
-
-/**
- * This hook is used to determine the position that a missing block should be inserted at.
- *
- * @return The index to insert the missing block at.
- */
-const findBlockPosition = ( {
-	defaultTemplatePosition,
-	innerBlocks,
-	currentDefaultTemplate,
-}: {
-	defaultTemplatePosition: number;
-	innerBlocks: BlockInstance[];
-	currentDefaultTemplate: MutableRefObject< TemplateArray >;
-} ) => {
-	switch ( defaultTemplatePosition ) {
-		case -1:
-			// The block is not part of the default template, so we append it to the current layout.
-			return innerBlocks.length;
-		// defaultTemplatePosition defaults to 0, so if this happens we can just return, this is because the block was
-		// the first block in the default layout, so we can prepend it to the current layout.
-		case 0:
-			return 0;
-		default:
-			// The new layout may have extra blocks compared to the default template, so rather than insert
-			// at the default position, we should append it after another default block.
-			const adjacentBlock =
-				currentDefaultTemplate.current[ defaultTemplatePosition - 1 ];
-			const position = innerBlocks.findIndex(
-				( { name: blockName } ) => blockName === adjacentBlock[ 0 ]
-			);
-			return position === -1 ? defaultTemplatePosition : position + 1;
-	}
-};
+import { getMissingBlocks, findBlockPosition } from './utils';
 
 /**
  * Hook to ensure FORCED blocks are rendered in the correct place.
@@ -112,11 +33,18 @@ export const useForcedLayout = ( {
 } ) => {
 	const currentRegisteredBlocks = useRef( registeredBlocks );
 	const currentDefaultTemplate = useRef( defaultTemplate );
-
 	const registry = useRegistry();
+	const { isPreview } = useEditorContext();
+
 	useEffect( () => {
 		let templateSynced = false;
+
+		if ( isPreview ) {
+			return;
+		}
+
 		const { replaceInnerBlocks } = dispatch( 'core/block-editor' );
+
 		return registry.subscribe( () => {
 			const innerBlocks = registry
 				.select( 'core/block-editor' )
@@ -181,5 +109,5 @@ export const useForcedLayout = ( {
 					.insertBlocks( blockConfig, insertAtPosition, clientId );
 			} );
 		}, 'core/block-editor' );
-	}, [ clientId, registry ] );
+	}, [ clientId, isPreview, registry ] );
 };
